Backend friends details

+++ backend/src/friends/friends.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateFriendDto } from './dto/create-friend.dto';
 import { UpdateFriendDto } from './dto/update-friend.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -17,6 +17,18 @@ export class FriendsService {
   ) {}
 
   public async create(createFriendDto: CreateFriendDto) {
+	  const found = await this.friendsRepository.find({
+		  where: [
+			  { requester: createFriendDto.requester, requested: createFriendDto.requested },
+			  { requester: createFriendDto.requested, requested: createFriendDto.requester },
+		  ],
+		});
+	if (found.length > 0) {
+		throw new ConflictException(['Friend request already exists.'], {
+			cause: new Error(),
+			description: `Friend request already exists.`,
+		});
+	}
     const friends: Friends = new Friends();
     friends.requester = createFriendDto.requester;
     friends.requested = createFriendDto.requested;
@@ -24,7 +36,57 @@ export class FriendsService {
     return await this.friendsRepository.save(friends);
   }
 
-  private getUsersData(friend: Friends, requested: Users, requester: Users) : FriendsResponse {
+  public async addFriendById(id: number, current_id: number) {
+	const found = await this.friendsRepository.find({
+		where: [
+			{ requester: current_id, requested: id },
+			{ requester: id, requested: current_id },
+		],
+	  });
+	if (found.length > 0) {
+		throw new ConflictException(['Friend request already exists.'], {
+			cause: new Error(),
+			description: `Friend request already exists.`,
+		});
+	}
+	const friends: Friends = new Friends();
+	friends.requester = current_id;
+	friends.requested = id;
+	friends.status = 0;
+	return await this.friendsRepository.save(friends);
+  }
+
+  public async addFriendByUsername(username: string, current_id: number) {
+	const found = await this.usersRepository.find({
+		where: { username: username },
+	  });
+	if (found.length == 0) {
+		throw new BadRequestException(['User not found.'], {
+			cause: new Error(),
+			description: `User not found.`,
+		});
+	}
+	const found2 = await this.friendsRepository.find({
+		where: [
+			{ requester: current_id, requested: found[0].id },
+			{ requester: found[0].id, requested: current_id },
+		],
+	  });
+	if (found2.length > 0) {
+		throw new ConflictException(['Friend request already exists.'], {
+			cause: new Error(),
+			description: `Friend request already exists.`,
+		});
+	}
+	const friends: Friends = new Friends();
+	friends.requester = current_id;
+	friends.requested = found[0].id;
+	friends.status = 0;
+	return await this.friendsRepository.save(friends);
+  }
+
+
+  private getUsersData(friend: Friends, requested: Users, requester: Users, count?: number) : FriendsResponse {
 	let tmp: FriendsResponse = {
 		id: friend.id,
 		requester: friend.requester,
@@ -39,15 +101,28 @@ export class FriendsService {
 			username: requested.username,
 			avatar: requested.avatar,
 			login42: requested.login42,
-		}
+		},
+		total_count: count
 	}
 	return tmp;
   }
 
-  public async findOne(id: number) {
+  public async findOne(id: number, user_id: number) {
 	const friends: Friends = await this.friendsRepository.findOne({
 	  where: { id: id },
 	});
+	if (friends == undefined) {
+		throw new NotFoundException(['Friend request not found.'], {
+			cause: new Error(),
+			description: `Friend request not found.`,
+		});
+	}
+	if (friends.requester != user_id && friends.requested != user_id) {
+		throw new BadRequestException(['Friend request not found.'], {
+			cause: new Error(),
+			description: `Friend request not found.`,
+		});
+	}
 	const users: Users[] = await this.usersRepository.find();
 	let requested: Users = users.find(user => {
 		return user.id == friends.requested
@@ -60,12 +135,24 @@ export class FriendsService {
   }
 
   public async findAllByUser(id: number) {
-	return await this.friendsRepository.find({
-	  where: [
-		{ requester: id },
-		{ requested: id },
-	  ],
+	const friendList: Friends[] =  await this.friendsRepository.find({
+		where: [
+			{ requester: id},
+			{ requested: id},
+		],
 	});
+	const users: Users[] = await this.usersRepository.find();
+	let ret: FriendsResponse[] = [];
+	friendList.forEach(friend => {
+		let requested: Users = users.find(user => {
+			return user.id == friend.requested
+		})
+		let requester : Users = users.find(user => {
+			return user.id == friend.requester
+		})
+		ret.push(this.getUsersData(friend, requested, requester, friendList.length));
+	})
+	return ret;
   }
 
   public async findFriendsList(id: number) {
@@ -84,9 +171,32 @@ export class FriendsService {
 		let requester : Users = users.find(user => {
 			return user.id == friend.requester
 		})
-		ret.push(this.getUsersData(friend, requested, requester));
+		ret.push(this.getUsersData(friend, requested, requester, friendlist.length));
 	})
 	return ret;
+  }
+
+  public async findFriendsListByUsername(username: string) {
+	const found = await this.usersRepository.find({
+		where: { username: username },
+	  });
+	if (found.length == 0) {
+		throw new BadRequestException(['User not found.'], {
+			cause: new Error(),
+			description: `User not found.`,
+		});
+	}
+	return this.findFriendsList(found[0].id);
+	}
+
+  public async findFriendsCount(id: number) {
+	const friendlist: Friends[] =  await this.friendsRepository.find({
+	  where: [
+		{ requester: id, status: 1 },
+		{ requested: id, status: 1 },
+	  ],
+	});
+	return friendlist.length;
   }
 
   public async findPendingFriends(id: number) {
@@ -102,10 +212,30 @@ export class FriendsService {
 		let requester: Users = users.find(user => {
 			return user.id == friend.requester
 		})
-		ret.push(this.getUsersData(friend, requested, requester));
+		ret.push(this.getUsersData(friend, requested, requester, pendinglist.length));
   })
   return ret;
 }
+
+public async findPendingFriendsByUsername(username: string) {
+	const found = await this.usersRepository.find({
+		where: { username: username },
+	  });
+	if (found.length == 0) {
+		throw new BadRequestException(['User not found.'], {
+			cause: new Error(),
+			description: `User not found.`,
+		});
+	}
+	return this.findPendingFriends(found[0].id);
+}
+
+  public async findPendingCount(id: number) {
+	const pendinglist: Friends[] = await this.friendsRepository.find({
+	  where: { requested: id, status: 0 },
+	});
+	return pendinglist.length;
+  }
 
   public async findPendingRequests(id: number) {
     const requestlist: Friends[] = await this.friendsRepository.find({
@@ -120,9 +250,29 @@ export class FriendsService {
 		let requester: Users = users.find(user => {
 			return user.id == friend.requester
 		})
-		ret.push(this.getUsersData(friend, requested, requester));
+		ret.push(this.getUsersData(friend, requested, requester, requestlist.length));
 	})
 	return ret;
+  }
+
+  public async findPendingRequestsByUsername(username: string) {
+	const found = await this.usersRepository.find({
+		where: { username: username },
+	  });
+	if (found.length == 0) {
+		throw new BadRequestException(['User not found.'], {
+			cause: new Error(),
+			description: `User not found.`,
+		});
+	}
+	return this.findPendingRequests(found[0].id);
+}
+
+  public async findRequestsCount(id: number) {
+	const requestlist: Friends[] = await this.friendsRepository.find({
+	  where: { requester: id, status: 0 },
+	});
+	return requestlist.length;
   }
 
   public async findAll() {
@@ -148,10 +298,72 @@ export class FriendsService {
     return await this.friendsRepository.update(id, friends);
   }
 
-  public async remove(id: number) {
+  public async remove(id: number, user_id: number) {
     const friends: Friends = await this.friendsRepository.findOne({
       where: { id: id },
     });
+	if (friends == undefined) {
+		throw new NotFoundException(['Friend request not found.'], {
+			cause: new Error(),
+			description: `Friend request not found.`,
+		});
+	}
+	if (friends.requester != user_id && friends.requested != user_id) {
+		throw new BadRequestException(['Friend request not found.'], {
+			cause: new Error(),
+			description: `Friend request not found.`,
+		});
+	}
     return await this.friendsRepository.remove(friends);
+  }
+
+  public async cancel(id: number, user_id: number) {
+	const friends: Friends = await this.friendsRepository.findOne({
+	  where: { id: id },
+	});
+	if (friends == undefined) {
+		throw new NotFoundException(['Friend request not found.'], {
+			cause: new Error(),
+			description: `Friend request not found.`,
+		});
+	}
+	if (friends.requester != user_id) {
+		throw new BadRequestException(['Friend request not found.'], {
+			cause: new Error(),
+			description: `Friend request not found.`,
+		});
+	}
+	if (friends.status != 0) {
+		throw new BadRequestException(['Friend request already accepted.'], {
+			cause: new Error(),
+			description: `Friend request already accepted.`,
+		});
+	}
+	return await this.friendsRepository.remove(friends);
+  }
+
+  public async reject(id: number, user_id: number) {
+	const friends: Friends = await this.friendsRepository.findOne({
+	  where: { id: id },
+	});
+	if (friends == undefined) {
+		throw new NotFoundException(['Friend request not found.'], {
+			cause: new Error(),
+			description: `Friend request not found.`,
+		});
+	}
+	if (friends.requested != user_id) {
+		throw new BadRequestException(['Friend request not found.'], {
+			cause: new Error(),
+			description: `Friend request not found.`,
+		});
+	}
+	if (friends.status != 0) {
+		throw new BadRequestException(['Friend request already accepted.'], {
+			cause: new Error(),
+			description: `Friend request already accepted.`,
+		});
+	}
+	return await this.friendsRepository.remove(friends);
   }
 }
